+++ app/api/v1/services/doctor_service.py
@@ -2,6 +2,7 @@ from typing import Any, Sequence
 
 from fastapi import HTTPException
 from passlib.context import CryptContext
+from sqlalchemy.exc import IntegrityError
 
 from app.api.v1.repositories.doctor_repository import DoctorRepository
 from app.schemas.schemas import DoctorRead, DoctorCreateRawPassword, DoctorCreateHashedPassword, \
@@ -23,16 +24,23 @@ class DoctorService:
         This method is used to retrieve all doctors from the DB.
 
         Returns:
-            doctors (Sequence[Patient])
+            doctors (Sequence[DoctorRead])
         """
 
         return await self.doctor_repository.get_doctors()
 
     async def get_doctor_by_id(self, doctor_id: int) -> DoctorRead:
-        """Retrieves a specific admin by ID."""
+        """
+        This method is used to retrieve a certain doctor from the DB by his 'id' field.
+
+        Returns:
+            doctor (DoctorRead | None)
+        """
+
         doctor = await self.doctor_repository.get_doctor_by_id(doctor_id)
         if not doctor:
-            raise HTTPException(status_code=404, detail="Doctor not found")
+            raise HTTPException(status_code=404, detail=f"Doctor with id {doctor_id} doest not exist.")
+
         return doctor
 
     async def create_doctor(self, raw_doctor_data: DoctorCreateRawPassword) -> dict[str, Any]:
@@ -41,7 +49,7 @@ class DoctorService:
         Moreover, this method hashes the raw password.
 
         Returns:
-            created doctor (dict[str, Any])
+            created doctor data (dict[str, Any])
         """
 
         hashed_password = hash_password(raw_doctor_data.password)
@@ -59,7 +67,7 @@ class DoctorService:
         This method is used to update the existing doctor data with the new one ('DoctorUpdateRawPassword' model).
 
         Returns:
-            updated doctor (dict[str, Any])
+            updated doctor (DoctorRead)
         """
 
         doctor_to_update = await self.doctor_repository.get_doctor_by_id(doctor_id)
@@ -76,12 +84,22 @@ class DoctorService:
 
         return await self.doctor_repository.update_doctor(doctor_with_hashed_password, doctor_id)
 
-    async def delete_doctor(self, doctor_id: int) -> int:
+    async def delete_doctor(self, doctor_id: int) -> dict:
         """
         This method is used to delete the existing doctor with given id.
 
         Returns:
-            deleted doctor ID (int)
+            A dictionary containing the deleted doctor ID and a message (dict)
         """
 
-        return await self.doctor_repository.delete_doctor(doctor_id)
+        doctor_to_delete = await self.doctor_repository.get_doctor_by_id(doctor_id)
+        if doctor_to_delete is None:
+            raise HTTPException(status_code=404, detail=f"Doctor with id {doctor_id} does not exist.")
+
+        try:
+            await self.doctor_repository.delete_doctor(doctor_id)
+            return {"doctor_id": doctor_id, "message": f"Doctor with id {doctor_id} has been deleted."}
+        except IntegrityError:
+            raise HTTPException(status_code=400,
+                                detail=f"Doctor with id {doctor_id} cannot be deleted because they have associated "
+                                       f"patients.")
